+++ tests/std_type_test.py
@@ -7,52 +7,52 @@ from monad_std import Option
 class TestStdList(unittest.TestCase):
     def test_index(self):
         x = MList([1, 2, 3, 4, 5])
-        self.assertEqual(x.index(2), Option.of_some(1))
-        self.assertEqual(x.index(0), Option.of_none())
-        self.assertEqual(x.get(2), Option.of_some(3))
-        self.assertEqual(x.get(10), Option.of_none())
+        self.assertEqual(x.index(2), Option.some(1))
+        self.assertEqual(x.index(0), Option.none())
+        self.assertEqual(x.get(2), Option.some(3))
+        self.assertEqual(x.get(10), Option.none())
 
     def test_pop(self):
         x = MList([1, 2, 3, 4, 5])
-        self.assertEqual(x.pop(), Option.of_some(5))
-        self.assertEqual(x.pop(1), Option.of_some(2))
+        self.assertEqual(x.pop(), Option.some(5))
+        self.assertEqual(x.pop(1), Option.some(2))
         x = MList()
-        self.assertEqual(x.pop(), Option.of_none())
+        self.assertEqual(x.pop(), Option.none())
 
 
 class TestStdTuple(unittest.TestCase):
     def test_index(self):
         x = MTuple((1, 3, "hey"))
-        self.assertEqual(x.index("hey"), Option.of_some(2))
-        self.assertEqual(x.index("hello"), Option.of_none())
-        self.assertEqual(x.get(1), Option.of_some(3))
-        self.assertEqual(x.get(5), Option.of_none())
+        self.assertEqual(x.index("hey"), Option.some(2))
+        self.assertEqual(x.index("hello"), Option.none())
+        self.assertEqual(x.get(1), Option.some(3))
+        self.assertEqual(x.get(5), Option.none())
 
 
 class TestStdDict(unittest.TestCase):
     def test_index(self):
         x = MDict({"a": 1, "b": 2, "c": 3})
-        self.assertEqual(x.get("b"), Option.of_some(2))
-        self.assertEqual(x.get("d"), Option.of_none())
+        self.assertEqual(x.get("b"), Option.some(2))
+        self.assertEqual(x.get("d"), Option.none())
 
     def test_pop(self):
         x = MDict({"a": 1})
-        self.assertEqual(x.popitem(), Option.of_some(("a", 1)))
-        self.assertEqual(x.popitem(), Option.of_none())
+        self.assertEqual(x.popitem(), Option.some(("a", 1)))
+        self.assertEqual(x.popitem(), Option.none())
 
         x = MDict({"a": 1})
-        self.assertEqual(x.pop("a"), Option.of_some(1))
-        self.assertEqual(x.pop("b"), Option.of_none())
-        self.assertEqual(x.pop("a"), Option.of_none())
+        self.assertEqual(x.pop("a"), Option.some(1))
+        self.assertEqual(x.pop("b"), Option.none())
+        self.assertEqual(x.pop("a"), Option.none())
 
 
 class TestStdSet(unittest.TestCase):
     def test_index(self):
         x = MSet([1, "d"])
         p1 = x.pop()
-        self.assertTrue(p1 == Option.of_some(1) or p1 == Option.of_some("d"))
+        self.assertTrue(p1 == Option.some(1) or p1 == Option.some("d"))
         x.pop()
-        self.assertEqual(x.pop(), Option.of_none())
+        self.assertEqual(x.pop(), Option.none())
 
 
 if __name__ == "__main__":